refactor(crawler): split monolithic crawler.py into src/crawler/ package
